Added remaining Cat functionality; Added User entity, controller, and service; Added Cat and User tests

+++ src/cats/cats.controller.ts
@@ -1,7 +1,8 @@
 import {
-  Body, Controller, Delete, Get, HttpException, HttpStatus, Param, Patch, Post, Put, Redirect,
+  Body, Controller, Delete, Get, HttpCode, HttpException, HttpStatus, Param, Patch, Post, Put, Redirect,
 } from '@nestjs/common';
-import { CreateOrUpdateCatDto } from './dto/create-or-update-cat.dto';
+import { CreateCatDto } from './dto/create-cat.dto';
+import { UpdateCatDto } from './dto/update-cat.dto';
 import { Cat } from './entities/cat.entity';
 import { CatsService } from './providers/cats.service';
 
@@ -12,6 +13,13 @@ export class CatsController {
   @Get()
   async findAll(): Promise<Cat[]> {
     return this.catsService.findAll();
+  }
+
+  @Get('random')
+  @Redirect()
+  async redirectToRandomCat(): Promise<{ url: string; statusCode: number; }> {
+    const id = await this.catsService.findRandomId();
+    return { url: `/cats/${id}`, statusCode: 302 };
   }
 
   @Get(':id')
@@ -25,40 +33,48 @@ export class CatsController {
   }
 
   @Post()
-  async create(@Body() dto: CreateOrUpdateCatDto): Promise<Cat> {
+  async create(@Body() dto: CreateCatDto): Promise<Cat> {
     const cat: Cat = new Cat();
     Object.assign(cat, dto);
     return this.catsService.save(cat);
   }
 
   @Put(':id')
-  async replace(@Param('id') id: number, @Body() dto: CreateOrUpdateCatDto) {
+  async replace(@Param('id') id: number, @Body() dto: UpdateCatDto): Promise<Cat> {
+    if (!(await this.catsService.exists(id))) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+
+    if (!Object.keys(dto).length) {
+      throw new HttpException('At least one valid property is required for an update.', HttpStatus.BAD_REQUEST);
+    }
+
     const cat: Cat = new Cat();
     Object.assign(cat, dto, { id });
-    this.catsService.save(cat);
+    await this.catsService.save(cat);
+    return this.catsService.find(id);
   }
 
   @Patch(':id')
-  async patch(@Param('id') id: number, @Body() dto: CreateOrUpdateCatDto) {
+  async patch(@Param('id') id: number, @Body() dto: UpdateCatDto): Promise<Cat> {
     const cat: Cat = await this.find(id);
     Object.assign(cat, dto);
-    this.catsService.save(cat);
+    return this.catsService.save(cat);
   }
 
   @Delete(':id')
-  async delete(@Param('id') id: number) {
+  @HttpCode(204)
+  async delete(@Param('id') id: number): Promise<void> {
+    if (!(await this.catsService.exists(id))) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+
     this.catsService.delete(id);
   }
 
   @Delete()
-  async deleteAll() {
+  @HttpCode(204)
+  async deleteAll(): Promise<void> {
     this.catsService.deleteAll();
-  }
-
-  @Get('random')
-  @Redirect()
-  async redirectToRandomCat() {
-    const id = await this.catsService.findRandomId();
-    return { url: `/cats/${id}`, statusCode: 302 };
   }
 }
